requirements.txt
# LangGraph + LangChain
langgraph
langchain
langchain-openai

# API
fastapi
uvicorn

# Database
supabase

# Dashboard
streamlit

# Utilities
pydantic
python-dotenv
faker
pandaspytest
